requirements.txt
discord_py_interactions==5.10.0
discord_py_slash_command==4.2.1
interactions.py==5.10.0
Requests==2.31.0
